rename some internal files
