+++ nexus/db-queries/src/db/datastore/mod.rs
@@ -46,6 +46,8 @@ use std::sync::Arc;
 
 mod address_lot;
 mod affinity;
+mod alert;
+mod alert_rx;
 mod allow_list;
 mod auth;
 mod bfd;
@@ -110,8 +112,6 @@ mod volume;
 mod volume_repair;
 mod vpc;
 pub mod webhook_delivery;
-mod webhook_event;
-mod webhook_rx;
 mod zpool;
 
 pub use address_lot::AddressLotCreateResult;
